+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Move to github CI
 - Add ASCT-B to matrix in README
 - Improve input validation logic
+- Use slice syntax to get list
 
 0.0.3
 - Change to more fluent SDK: `select_TARGET(where='SOURCE', has='CRITERIA', ...)`